Give MAXPATHL, IOSIZE and READBIN one home each
`MAXPATHL` was re-declared in 42 files (plus `DEFAULT_MAXPATHL` in 15),
`IOSIZE` in 38, `READBIN` in 4. Values agreed everywhere: 4096, 1025,
`c"rb"`. Canonical homes follow the upstream headers -- `types::os` for
`os_defs.h`'s two path constants and `macros_defs.h`'s `READBIN`,
`types::globals` for `globals.h`'s `IOSIZE`.

Both size constants are `c_int`, which is what the majority of the copies
were and what the tree already assumes: `MAXPATHL as usize` appeared 76
times and `IOSIZE as usize` 17 times before this commit, so the ~90 new
casts join an established idiom rather than inventing one.

One value mismatch, and it was a naming bug rather than a divergence:
`eval/system.rs` called its 1025 `MAXPATHL`. Upstream's `tv_to_argv`
declares that buffer `char buf[IOSIZE]`, so the constant was right and
the name was wrong; it is now `IOSIZE`.

`--allow-growth`: +30 unchecked lines. Every one is rustfmt reflowing a
call that grew by ` as usize`; no new unsafe operation appears.

+++ crates/nvim/src/file_search/init.rs
@@ -12,7 +12,7 @@
 
 use super::*;
 use crate::semsg_c;
-use crate::types::FAIL;
+use crate::types::{FAIL, MAXPATHL};
 use ::libc::strtol;
 use core::ffi::{c_char, c_int, c_void};
 use core::{ptr, slice};
@@ -75,20 +75,22 @@ unsafe fn starting_dir(
             && !rel_fname.is_null()
         {
             let len = path_tail(rel_fname.cast_mut()).offset_from(rel_fname) as usize;
-            ctx.start_dir = Some(if !vim_isAbsName(rel_fname) && len + 1 < MAXPATHL {
-                // Make the start dir an absolute path name.
-                full_name_of(name_of(rel_fname, len).as_ptr(), false)
-            } else {
-                name_of(rel_fname, len)
-            });
+            ctx.start_dir = Some(
+                if !vim_isAbsName(rel_fname) && len + 1 < MAXPATHL as usize {
+                    // Make the start dir an absolute path name.
+                    full_name_of(name_of(rel_fname, len).as_ptr(), false)
+                } else {
+                    name_of(rel_fname, len)
+                },
+            );
             // Step over the "." and the separator after it, if any.
             let path = path.add(1);
             return Ok(if *path != 0 { path.add(1) } else { path });
         }
 
         if *path == 0 || !vim_isAbsName(path) {
-            let mut curdir = [0 as c_char; MAXPATHL];
-            if os_dirname(curdir.as_mut_ptr(), MAXPATHL) == FAIL {
+            let mut curdir = [0 as c_char; MAXPATHL as usize];
+            if os_dirname(curdir.as_mut_ptr(), MAXPATHL as usize) == FAIL {
                 return Err(());
             }
             ctx.start_dir = Some(Name::from_ptr(curdir.as_ptr()));
@@ -122,7 +124,7 @@ unsafe fn stop_directories(stopdirs: *mut c_char) -> Vec<Name> {
             };
 
             dirs.push(
-                if *entry != 0 && !vim_isAbsName(entry) && len + 1 < MAXPATHL {
+                if *entry != 0 && !vim_isAbsName(entry) && len + 1 < MAXPATHL as usize {
                     // Upstream copies the entry into a scratch buffer and then
                     // resolves `entry` instead, which is not NUL-terminated at
                     // the ';': a relative stop directory with another after it
@@ -158,7 +160,7 @@ unsafe fn wildcard_tail(wc_part: *mut c_char) -> Result<Name, ()> {
         let mut tail = Vec::<u8>::new();
         let mut at = wc_part;
         while *at != 0 {
-            if tail.len() + 5 >= MAXPATHL {
+            if tail.len() + 5 >= MAXPATHL as usize {
                 emsg(gettext(c"E854: Path too long for completion".as_ptr()));
                 break;
             }
@@ -210,7 +212,7 @@ unsafe fn first_frame(ctx: &mut FindContext) -> Result<Name, ()> {
     unsafe {
         let start_dir = ctx.start_dir.as_ref().expect("set above");
         // Create an absolute path.
-        if start_dir.len() + ctx.fix_path.len() + 3 >= MAXPATHL {
+        if start_dir.len() + ctx.fix_path.len() + 3 >= MAXPATHL as usize {
             emsg(gettext(c"E854: Path too long for completion".as_ptr()));
             return Err(());
         }
