+++ crates/nvim/src/os/env/dirs.rs
@@ -16,7 +16,7 @@ use crate::memory::xmemrchr;
 use crate::os::fs::os_isdir;
 use crate::path::{after_pathsep, append_path, concat_fnames, path_fnamencmp, path_tail_with_sep};
 use crate::strings::vim_strchr;
-use crate::types::{OK, VV_PROGPATH, buf_T};
+use crate::types::{MAXPATHL, OK, VV_PROGPATH, buf_T};
 
 /// The directory a runtime lives in, under `$VIM`.
 const RUNTIME_DIRNAME: &CStr = c"runtime";
@@ -144,7 +144,7 @@ pub unsafe fn vim_get_prefix_from_exepath(exe_name: *mut c_char) {
     // SAFETY: the caller's contract; `path_tail*` answer pointers inside the
     // buffer they are given.
     unsafe {
-        xstrlcpy(exe_name, get_vim_var_str(VV_PROGPATH), MAXPATHL);
+        xstrlcpy(exe_name, get_vim_var_str(VV_PROGPATH), MAXPATHL as usize);
         // Remove the trailing "nvim", then the trailing "bin/".
         *path_tail_with_sep(exe_name) = 0;
         *path_tail(exe_name) = 0;
@@ -194,7 +194,7 @@ pub unsafe fn vim_getenv(name: *const c_char) -> *mut c_char {
 
         // Still nothing: work backwards from 'helpfile' (unless it holds a
         // '$'), then from the executable's own path.
-        let mut exe_name: [c_char; MAXPATHL] = [0; MAXPATHL];
+        let mut exe_name: [c_char; MAXPATHL as usize] = [0; MAXPATHL as usize];
         if vim_path.is_null() {
             let from_helpfile =
                 !p_hf.get().is_null() && vim_strchr(p_hf.get(), '$' as c_int).is_null();
@@ -206,7 +206,7 @@ pub unsafe fn vim_getenv(name: *const c_char) -> *mut c_char {
                 if append_path(
                     exe_name.as_mut_ptr(),
                     c"share/nvim/runtime/".as_ptr(),
-                    MAXPATHL,
+                    MAXPATHL as usize,
                 ) == OK
                 {
                     vim_path = exe_name.as_mut_ptr();
